// webui/src/lib/gen/models/BodyLoginLoginPost.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */

export interface BodyLoginLoginPost {
	grant_type?: string | null;
	username: string;
	password: string;
	scope?: string;
	client_id?: string | null;
	client_secret?: string | null;
}