// webui/src/lib/gen/models/bodyLoginLoginPost.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { BodyLoginLoginPostGrantType } from './bodyLoginLoginPostGrantType';
import type { BodyLoginLoginPostClientId } from './bodyLoginLoginPostClientId';
import type { BodyLoginLoginPostClientSecret } from './bodyLoginLoginPostClientSecret';

export interface BodyLoginLoginPost {
	grant_type?: BodyLoginLoginPostGrantType;
	username: string;
	password: string;
	scope?: string;
	client_id?: BodyLoginLoginPostClientId;
	client_secret?: BodyLoginLoginPostClientSecret;
}